Move blank node linkers to whelk-core
