+++ whelk-core/src/main/groovy/whelk/util/Statistics.groovy
@@ -1,4 +1,4 @@
-package datatool.util
+package whelk.util
 
 import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.atomic.AtomicInteger
